feat(ProcessNoteRange, ProcessPitch): move folders
